% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.conflicts.OK}
\alias{.conflicts.OK}
\title{Suppress conflict warnings}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.conflicts.OK
}
\description{
Set to \code{TRUE} in the package environment during \code{.onAttach()} so that
\code{base::library()} skips its default "masked objects" messages.
}
\keyword{internal}
